fix(auth): move PasskeyRegisterExternalView out of page file for Next.js build
Next.js 15 page.tsx files only allow default + specific metadata
exports. Exporting PasskeyRegisterExternalView and its Status type from
the page broke production build with 'not a valid Page export field'.

Extract the pure view component and its Status type into
apps/web/src/components/auth/PasskeyRegisterExternalView.tsx. The page
now only exports the default and imports the view from its new home.
Tests updated to import from the component module.
